Only dropping db locally

+++ python-backend/env_setup.py
@@ -2,8 +2,9 @@ import sys_modules.database as database
 
 from pymongo import MongoClient
 
-c = MongoClient()
-c.drop_database('community')
+if database.uri:
+  c = MongoClient()
+  c.drop_database('community')
 
 # Insert admin user
 database.db_insert( 'users', {'username': 'admin'} )
